src/backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-jose
python-multipart
passlib==1.7.4
bcrypt==4.0.1
cryptography
pydantic
supabase
